% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractiveManhattanPlotter.R
\name{InteractiveManhattanPlotter}
\alias{InteractiveManhattanPlotter}
\title{Plot an interactive Manhattan plot and save it as an HTML}
\usage{
InteractiveManhattanPlotter(
  gwas_result_file = "data/example/FHS_EA_MRS_merged.txt",
  chr = "CHR",
  pos = "POS",
  ref = "NEA",
  alt = "EA",
  pvalue = "p.value",
  delimiter = " ",
  color1 = "cornflowerblue",
  color2 = "blue4",
  threshold_linecolor = "black"
)
}
\arguments{
\item{gwas_result_file}{The name of the text file containing GWAS statistics}

\item{chr}{The name of the chromosome number column in the text file}

\item{pos}{The name of the chromosome position column in the text file}

\item{ref}{The name of the reference allele column in the text file}

\item{alt}{The name of the alternative allele column in the text file}

\item{pvalue}{The name of the pvalue column in the text file}

\item{delimiter}{The delimiter used in the text file}

\item{color1}{First color of the Manhattan plot}

\item{color2}{Second color of the Manhattan plot}

\item{threshold_linecolor}{Default is set as black}
}
\value{
A html widget of the Manhattan plot generated
}
\description{
Plot an interactive Manhattan plot and save it as an HTML
}
\examples{
InteractiveManhattanPlotter("data/example/FHS_EA_MRS_merged.txt")
}
