% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayDiseaseAnnotator.R
\name{PathwayDiseaseAnnotator}
\alias{PathwayDiseaseAnnotator}
\title{Annotate the GWAS results and include pathway and disease information}
\usage{
PathwayDiseaseAnnotator(
  variant_file = "data/example/FHS_EA_MRS_5e8_snplist.txt",
  chr = "CHR",
  pos = "POS",
  ref = "NEA",
  alt = "EA",
  delimiter = " ",
  write_excel = TRUE
)
}
\arguments{
\item{variant_file}{The name of the text file containing GWAS statistics}

\item{chr}{The name of the chromosome number column in the text file}

\item{pos}{The name of the chromosome position column in the text file}

\item{ref}{The name of the reference allele column in the text file}

\item{alt}{The name of the alternative allele column in the text file}

\item{delimiter}{The delimiter used in the text file}

\item{write_excel}{Whether to write the result to an excel file, else write as CSV}

\item{pvalue}{The name of the pvalue column in the text file}
}
\value{
A data frame of the annotated results
}
\description{
Annotate the GWAS results and include pathway and disease information
}
\examples{
PathwayDiseaseAnnotator("data/example/FHS_EA_MRS_merged.txt")
}
